fix: 半休が切り捨てられていたバグを修正

+++ WEB-INF/src/fukaisystem/application/attendance/ReadAll.java
@@ -117,7 +117,7 @@ public class ReadAll extends ServiceFoundation {
 				v.add(rs.getInt("日曜"));
 				v.add(rs.getInt("出張"));
 				v.add(rs.getInt("欠勤"));
-				v.add(rs.getInt("有給"));
+				v.add(rs.getFloat("有給"));
 				dataVector.add(v);
 			}
 		}
